split tree.py into multiple files.
